pyproject.toml
[project]
name = "archives"
version = "0.1.0"
description = "microservice for media uploaded files for cema"
authors = [
    {name = "Muchiri Antony"}
]
license = {text = "MIT"}
requires-python = "3.14"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
